docs(site): Split functions into separate files
