+++ docs/src/components/App/Header/Header.less
@@ -140,13 +140,3 @@
     margin-bottom: @row-height;
   }
 }
-
-.highlightWrapper {
-  position: relative;
-}
-
-.highlightedIcon {
-  position: absolute;
-  top: 16px;
-  left: -24px;
-}
