+++ docs/src/components/App/Header/WithHighlighting/WithHighlighting.less
@@ -1,0 +1,9 @@
+.highlightWrapper {
+  position: relative;
+}
+
+.highlightedIcon {
+  position: absolute;
+  top: 16px;
+  left: -24px;
+}
